+++ docs/web_modules/svelte.js
@@ -1,1 +1,1 @@
-export { g as getContext, o as onDestroy, a as onMount, s as setContext } from './common/index-a391cc2f.js';
+export { g as getContext, o as onDestroy, a as onMount, s as setContext, t as tick } from './common/index-e2711807.js';